Add the async() helper and drone
The async() helper inserts a placeholder in to the document and
will execute the function you pass to it.

That function has another callback as argument, one you must call
again with the result to replace the placeholder with.

It can directly  be used inside templates if you like, but it's
much better suited for use inside other helpers.